refactor: extract audit-lock helpers to AuditChainLockTrait
SonarCloud S1448 (max methods per class) flagged AuditLogService at
23 methods / 20 allowed after the four lock helpers landed.

Moving the four `acquireAuditLock` / `commit` / `rollback` /
`releaseAuditLock` helpers into a dedicated `AuditChainLockTrait`
drops AuditLogService back under the threshold while keeping the
runtime audit-write path readable.

The trait deliberately stays runtime-specific (throws
`AuditWriteException` on lock failure). Migration paths
(`AuditHmacMigrationWizard` + `VaultAuditMigrateCommand`) inline the
same shape but throw `AuditMigrationException` — sharing the trait
between contexts would erase the runtime-vs-migration distinction
surfaced in PR #133 review. Worth revisiting only if a third caller
appears.

No behaviour change. 1741 unit tests pass; cgl + rector green.

Refs SonarCloud S1448 on PR #134.
